fix: move classes to utils
